v2.8.4.1

+++ 140_dbs/src/gplx/dbs/sqls/Db_fld.java
@@ -17,7 +17,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 package gplx.dbs.sqls; import gplx.*; import gplx.dbs.*;
 public class Db_fld {
-	public Db_fld(String name, byte type_tid) {this.name = name; this.type_tid = type_tid;}
+	public Db_fld(String name, int type_tid) {this.name = name; this.type_tid = type_tid;}
 	public String Name() {return name;} public Db_fld Name_(String v) {name = v; return this;} private String name;
-	public byte Type_tid() {return type_tid;} public Db_fld Type_tid_(byte v) {type_tid = v; return this;} private byte type_tid;
+	public int Type_tid() {return type_tid;} public Db_fld Type_tid_(int v) {type_tid = v; return this;} private int type_tid;
 }
